FIX AUTHORIZATION

+++ siswa/siswa.php
@@ -13,6 +13,16 @@ require_once "../template/header.php";
 require_once "../template/navbar.php";
 require_once "../template/sidebar.php";
 
+$username = $_SESSION["ssUser"];
+$queryUser = mysqli_query($koneksi, "SELECT * FROM user WHERE username = '$username'");
+$profile = mysqli_fetch_array($queryUser);
+
+if ($profile['jabatan'] == "Guru") {
+    $displayGuru = "visually-hidden";
+} else {
+    $displayGuru = "";
+}
+
 ?>
 
 <div id="layoutSidenav_content">
@@ -27,8 +37,7 @@ require_once "../template/sidebar.php";
             <div class="card">
                 <div class="card-header">
                     <span class="h5 my-2"><i class="fa-solid fa-list"></i> Data Siswa</span>
-                    <a href="<?= $main_url ?>siswa/add-siswa.php" class="btn btn-sm btn-primary float-end"><i
-                            class="fa-solid fa-plus"></i> Tambah Siswa</a>
+                    <a href="<?= $main_url ?>siswa/add-siswa.php" class="btn btn-sm btn-primary float-end <?= $displayGuru ?>"><i class="fa-solid fa-plus"></i> Tambah Siswa</a>
                 </div>
                 <div class="card-body">
                     <table class="table table-hover" id="datatablesSimple">
@@ -54,7 +63,7 @@ require_once "../template/sidebar.php";
                                     <center>Alamat</center>
                                 </th>
                                 <th scope="col">
-                                    <center>Operasi</center>
+                                    <center class="<?= $displayGuru ?>">Operasi</center>
                                 </th>
                             </tr>
                         </thead>
@@ -64,29 +73,24 @@ require_once "../template/sidebar.php";
                             $querySiswa = mysqli_query($koneksi, "SELECT * FROM siswa");
 
                             while ($data = mysqli_fetch_array($querySiswa)) { ?>
-                            <tr>
-                                <th scope="row"><?= $no++ ?></th>
-                                <td align="center">
-                                    <center>
-                                        </cente><img src="../asset/image/<?= $data['foto'] ?>" class="rounded-circle"
-                                            alt="Foto Siswa" width="60px"></center>
-                                </td>
-                                <td><?= $data['nis'] ?></td>
-                                <td><?= $data['nama'] ?></td>
-                                <td><?= $data['kelas'] ?></td>
-                                <td><?= $data['jurusan'] ?></td>
-                                <td><?= $data['alamat'] ?></td>
-                                <td align="center">
-                                    <center>
-                                        <a href="edit-siswa.php?nis=<?=$data['nis']?>" class="btn btn-sm btn-warning"
-                                            title="Update Siswa"><i class="fa-solid fa-pen"></i></a>
-                                        <a href="hapus-siswa.php?nis=<?= $data['nis'] ?>&foto=<?= $data['foto'] ?>"
-                                            class="btn btn-sm btn-danger" title="Hapus Siswa"
-                                            onclick="return confirm('Anda yakin akan menghapus data ini ?')"><i
-                                                class="fa-solid fa-trash"></i>
-                                    </center></a>
-                                </td>
-                            </tr>
+                                <tr>
+                                    <th scope="row"><?= $no++ ?></th>
+                                    <td align="center">
+                                        <center>
+                                            </cente><img src="../asset/image/<?= $data['foto'] ?>" class="rounded-circle" alt="Foto Siswa" width="60px"></center>
+                                    </td>
+                                    <td><?= $data['nis'] ?></td>
+                                    <td><?= $data['nama'] ?></td>
+                                    <td><?= $data['kelas'] ?></td>
+                                    <td><?= $data['jurusan'] ?></td>
+                                    <td><?= $data['alamat'] ?></td>
+                                    <td align="center">
+                                        <center>
+                                            <a href="edit-siswa.php?nis=<?= $data['nis'] ?>" class="btn btn-sm btn-warning <?= $displayGuru ?>" title="Update Siswa"><i class="fa-solid fa-pen"></i></a>
+                                            <a href="hapus-siswa.php?nis=<?= $data['nis'] ?>&foto=<?= $data['foto'] ?>" class="btn btn-sm btn-danger <?= $displayGuru ?>" title="Hapus Siswa" onclick="return confirm('Anda yakin akan menghapus data ini ?')"><i class="fa-solid fa-trash"></i>
+                                        </center></a>
+                                    </td>
+                                </tr>
                             <?php } ?>
                         </tbody>
                     </table>
